backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-dotenv==1.0.1
pydantic==2.6.3
pydantic-settings==2.2.1
pydantic[email]==2.6.3
python-multipart==0.0.9
faker==24.4.0
email-validator==2.1.1
